Add FixmeStatus to Fixme struct
This adds a enum consisting of Active and Complete to the Fixme
struct. When viewing and completing fixmes, having a indication of
whether this fixme has already been done or not is essential.

"Completing" fixmes *could* have been accomplished by simply removing
them from the configuration file, but being able to see completed
fixmes could be helpful as well. Additionally this could pave the way
for an "undo" command later on.

<!-- ps-id: cb9ecf90-0afe-4c57-958a-023301c20f24 -->