Refactor commands::list::list to use IndexedFixme struct.
The `IndexedFixme` struct was built to store information about the
identifiers of a fixme and its project. Simply calling
`.iter().enumerate()` gives us such an identifier for both. And
because `list` now returns a `Vec` of `IndexedFixme`, we can simply
use the `Display` trait implemented on it, which allows us to clean up
the main loop.

<!-- ps-id: cba2b57a-4f1a-444a-8209-285d6e226c87 -->