Clean up fixme listing display
There were many explicit UI elements that made the display overly busy
and difficult to parse at a glance. This removes many of them but
wraps different elements in some form of bracketing.

Additionally, the hyphen between the project id and fixme id has been
removed to mitigate confusion. When fixing or deleting a fixme, the
project id and the fixme id need to be specified space-delimited, but
the former UI presented them hyphen-delimited.

<!-- ps-id: b4303691-154b-46fd-ba26-85cd0e4e8337 -->